requirements.txt
requests
Pillow
beautifulsoup4
lxml
cairosvg
rembg==2.0.56
